rename OpenAPI- to JwtAuthConvention
