+++ server-scala/stateless-openapi/src/main/scala/io/jokester/fullstack_playground/user_todolist_api/UserTodoApi.scala
@@ -1,7 +1,7 @@
 package io.jokester.fullstack_playground.user_todolist_api
 
 import io.circe.generic.auto._
-import io.jokester.http_api.OpenAPIAuthConvention._
+import io.jokester.http_api.JwtAuthConvention._
 import io.jokester.http_api.OpenAPIBuilder
 import io.jokester.http_api.OpenAPIConvention._
 import sttp.tapir.generic.auto._
